--- pkg/cm_accounts/cm_account.go
@@ -1,95 +1,0 @@
-// Copyright (C) 2022-2025, Chain4Travel AG. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package cmaccounts
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"github.com/chain4travel/camino-messenger-contracts/go/contracts/cmaccount"
-	"github.com/chain4travel/camino-messenger-contracts/go/contracts/cmaccountmanager"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/ethclient"
-	"go.uber.org/zap"
-)
-
-type CmAccountService interface {
-	WarnIfUpgradeNeeded() error
-}
-
-type cmAccountService struct {
-	cmAccountAddress *common.Address
-	cmAccount        *cmaccount.Cmaccount
-	manager          *cmaccountmanager.Cmaccountmanager
-	logger           *zap.SugaredLogger
-	ethClient        *ethclient.Client
-}
-
-func NewCmAccountService(
-	cmAccountAddress common.Address,
-	logger *zap.SugaredLogger,
-	ethClient *ethclient.Client,
-) (CmAccountService, error) {
-	cmAccount, err := cmaccount.NewCmaccount(cmAccountAddress, ethClient)
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch CM account: %w", err)
-	}
-
-	managerAddress, err := cmAccount.GetManagerAddress(&bind.CallOpts{})
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch CM account Manager Address: %w", err)
-	}
-	manager, err := cmaccountmanager.NewCmaccountmanager(managerAddress, ethClient)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get Manager: %w", err)
-	}
-
-	return &cmAccountService{
-		cmAccountAddress: &cmAccountAddress,
-		cmAccount:        cmAccount,
-		manager:          manager,
-		logger:           logger,
-		ethClient:        ethClient,
-	}, nil
-}
-
-func (s *cmAccountService) WarnIfUpgradeNeeded() error {
-	currentImplOnManager, err := s.manager.GetAccountImplementation(&bind.CallOpts{})
-	if err != nil {
-		return fmt.Errorf("failed to get Account Implementation: %w", err)
-	}
-
-	// Implementation slot for ERC1967Proxy
-	// See: https://eips.ethereum.org/EIPS/eip-1967#logic-contract-address
-	const implementationSlotString = "0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc"
-	implementationSlot := common.HexToHash(implementationSlotString)
-	// Read implementation from proxy
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-	defer cancel()
-	implAddress, err := s.ethClient.StorageAt(ctx, *s.cmAccountAddress, implementationSlot, nil)
-	if err != nil {
-		return fmt.Errorf("failed to get implementation address from proxy: %w", err)
-	}
-
-	if len(implAddress) < 32 {
-		return fmt.Errorf("implementation address storage read returned unexpected size: %d", len(implAddress))
-	}
-	currentImplOnProxy := common.BytesToAddress(implAddress[12:])
-
-	s.logger.Info("Implementation:")
-	s.logger.Info("   - Active:  " + currentImplOnProxy.Hex())
-	s.logger.Info("   - Latest:  " + currentImplOnManager.Hex())
-
-	if currentImplOnProxy != currentImplOnManager {
-		// TODO: @VjeraTurk Ensure multiple versions compatibility
-		// TODO: @VjeraTurk Inform about the consequences of not upgrading for specific cases (bookingtoken, manager, cmaccount ...)
-		s.logger.Error("CMAccount needs an upgrade!")
-	} else {
-		s.logger.Info("CMAccount is using the latest implementation.")
-	}
-	return nil
-}
